+++ internal/downloading/dumper.go
@@ -77,7 +77,7 @@ func (td *TweetDumper) Dump(path string) error {
 	if err != nil {
 		return err
 	}
-	return os.WriteFile(path, data, 0666)
+	return os.WriteFile(path, data, 0600)
 }
 
 func (td *TweetDumper) Clear() {